Refactorizar acordiones de stdout y stderr en modal de resultados
- Ambos componentes eran parecidos por lo que se decidió compartir el
código. Ahora se usa StdAccordion para ambos casos.
